fix type missmatch

+++ NEK/Nikon.cpp
@@ -266,7 +266,7 @@ mtp::MtpDevicePropDescDS NikonCamera::GetDevicePropDesc(uint32_t devicePropCode)
 
 	return mtp::MtpDevice::GetDevicePropDesc(devicePropCode);
 }
-mtp::MtpDatatypeVariant NikonCamera::GetDevicePropValue(uint16_t devicePropCode) {
+mtp::MtpDatatypeVariant NikonCamera::GetDevicePropValue(uint32_t devicePropCode) {
 	mutexDeviceInfo_.lock();
 	if (std::find(deviceInfo_.OperationsSupported.begin(), deviceInfo_.OperationsSupported.end(), NikonMtpOperationCode::GetDevicePropValueEx) != deviceInfo_.OperationsSupported.end()) {
 		mutexDeviceInfo_.unlock();
@@ -279,7 +279,7 @@ mtp::MtpDatatypeVariant NikonCamera::GetDevicePropValue(uint16_t devicePropCode)
 		mutexDeviceInfo_.unlock();
 
 		mtp::MtpParams params;
-		params.addUint32(static_cast<uint32_t>(devicePropCode));
+		params.addUint32(devicePropCode);
 		mtp::MtpResponse response = SendCommandAndRead(NikonMtpOperationCode::GetDevicePropValueEx, params);
 
 		if (response.responseCode != mtp::MtpResponseCode::OK) {
@@ -292,14 +292,14 @@ mtp::MtpDatatypeVariant NikonCamera::GetDevicePropValue(uint16_t devicePropCode)
 
 	return mtp::MtpDevice::GetDevicePropValue(devicePropCode);
 }
-void NikonCamera::SetDevicePropValue(uint16_t devicePropCode, mtp::MtpDatatypeVariant data) {
+void NikonCamera::SetDevicePropValue(uint32_t devicePropCode, mtp::MtpDatatypeVariant data) {
 	mutexDeviceInfo_.lock();
 	if (std::find(deviceInfo_.OperationsSupported.begin(), deviceInfo_.OperationsSupported.end(), NikonMtpOperationCode::SetDevicePropValueEx) != deviceInfo_.OperationsSupported.end()) {
 		mutexDeviceInfo_.unlock();
 		std::vector<uint8_t> rawdata = SetDevicePropValue_(data);
 
 		mtp::MtpParams params;
-		params.addUint32(static_cast<uint32_t>(devicePropCode));
+		params.addUint32(devicePropCode);
 		mtp::MtpResponse response = SendCommandAndWrite(NikonMtpOperationCode::SetDevicePropValueEx, params, rawdata);
 
 		if (response.responseCode != mtp::MtpResponseCode::OK) {
